+++ src/boot.cpp
@@ -33,7 +33,7 @@ extern "C" {
 extern const uint8_t BOOTIMAGE_SYMBOL(start)[];
 extern const uint8_t BOOTIMAGE_SYMBOL(end)[];
 
-AVIAN_EXPORT const uint8_t* bootimageBin(unsigned* size)
+AVIAN_EXPORT const uint8_t* bootimageBin(size_t* size)
 {
   *size = BOOTIMAGE_SYMBOL(end) - BOOTIMAGE_SYMBOL(start);
   return BOOTIMAGE_SYMBOL(start);
@@ -42,7 +42,7 @@ AVIAN_EXPORT const uint8_t* bootimageBin(unsigned* size)
 extern const uint8_t CODEIMAGE_SYMBOL(start)[];
 extern const uint8_t CODEIMAGE_SYMBOL(end)[];
 
-AVIAN_EXPORT const uint8_t* codeimageBin(unsigned* size)
+AVIAN_EXPORT const uint8_t* codeimageBin(size_t* size)
 {
   *size = CODEIMAGE_SYMBOL(end) - CODEIMAGE_SYMBOL(start);
   return CODEIMAGE_SYMBOL(start);
@@ -65,7 +65,7 @@ extern "C" {
 extern const uint8_t SYMBOL(start)[];
 extern const uint8_t SYMBOL(end)[];
 
-AVIAN_EXPORT const uint8_t* classpathJar(unsigned* size)
+AVIAN_EXPORT const uint8_t* classpathJar(size_t* size)
 {
   *size = SYMBOL(end) - SYMBOL(start);
   return SYMBOL(start);
